Initial commit of the new onboarding questions in the welcome flow
